// src/lib/strapi/types/api/lien-rex-diagnostic.ts
// Interface automatically generated by schemas-to-ts

import { FicheDiagnostic } from "./fiche-diagnostic";
import { RetourExperienceDiagnostic } from "./retour-experience-diagnostic";
import { FicheDiagnostic_Plain } from "./fiche-diagnostic";
import { RetourExperienceDiagnostic_Plain } from "./retour-experience-diagnostic";
import { AdminPanelRelationPropertyModification } from "../common/AdminPanelRelationPropertyModification";

export interface LienRexDiagnostic {
  id: number;
  attributes: {
    createdAt: Date;
    updatedAt: Date;
    publishedAt?: Date;
    fiche_diagnostic?: { data: FicheDiagnostic };
    retour_experience_diagnostic?: { data: RetourExperienceDiagnostic };
    description?: any;
  };
}
export interface LienRexDiagnostic_Plain {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  fiche_diagnostic?: FicheDiagnostic_Plain;
  retour_experience_diagnostic?: RetourExperienceDiagnostic_Plain;
  description?: any;
}

export interface LienRexDiagnostic_NoRelations {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  fiche_diagnostic?: number;
  retour_experience_diagnostic?: number;
  description?: any;
}

export interface LienRexDiagnostic_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  fiche_diagnostic?: AdminPanelRelationPropertyModification<FicheDiagnostic_Plain>;
  retour_experience_diagnostic?: AdminPanelRelationPropertyModification<RetourExperienceDiagnostic_Plain>;
  description?: any;
}